LPS-153063 Rename reducers and move them to corresponding folder
