+++ modules/apps/style-book/style-book-web/src/main/resources/META-INF/resources/js/style-book-editor/LayoutContext.js
@@ -19,7 +19,7 @@ import {
 	SET_PREVIEW_LAYOUT,
 	SET_PREVIEW_LAYOUT_TYPE,
 } from './constants/actionTypes';
-import layoutReducer from './layoutReducer';
+import layoutReducer from './reducers/layoutReducer';
 
 const LayoutDispatchContext = React.createContext(() => {});
 export const LayoutStoreContext = React.createContext({
