+++ modules/apps/style-book/style-book-web/src/main/resources/META-INF/resources/js/style-book-editor/reducers/layoutReducer.js
@@ -16,7 +16,7 @@ import {
 	LOADING,
 	SET_PREVIEW_LAYOUT,
 	SET_PREVIEW_LAYOUT_TYPE,
-} from './constants/actionTypes';
+} from '../constants/actionTypes';
 
 export default function layoutReducer(state, action) {
 	switch (action.type) {
